Got tree saving/recoupling from the database in unit tests.

+++ index.js
@@ -1,4 +1,5 @@
-var mongo = require('promised-mongo');
+var mongo = require('promised-mongo'),
+	Q = require('q');
 
 var Adapter = function(settings) {
 	this.settings = settings || {};
@@ -15,11 +16,37 @@ Adapter.prototype = {
 		this.db = undefined;
 	},
 	find: function(treeId) {
-		return this.db[this.settings.treeCollection].find({
+		if (!this.db)
+			throw Error('Need to connect!');
+
+		var that = this;
+		var tree = undefined;
+
+		var deferred = Q.defer();
+
+		var promise = this.db[this.settings.treeCollection].find({
 			'_id': treeId
+		}).toArray().then(function(t) {
+			tree = t[0];
+			return that.db[that.settings.nodeCollection].find({
+				'_id': {
+					$in: tree.nodeIds
+				}
+			}).toArray().then(function(nodes) {
+				deferred.resolve({
+					tree: tree,
+					nodes: nodes
+				});
+			});
+		}, function(err) {
+			deferred.reject(err);
 		});
+
+		return deferred.promise;
 	},
 	insert: function(tree) {
+		if (!this.db)
+			throw Error('Need to connect!');
 		var treeDoc = tree.decouple();
 		var that = this;
 		return this.db[this.settings.treeCollection].insert(treeDoc.tree).then(function() {
@@ -29,13 +56,16 @@ Adapter.prototype = {
 		});
 	},
 	drop: function() {
+		if (!this.db)
+			throw Error('Need to connect!');
 		var that = this;
 		return this.db[this.settings.treeCollection].drop().then(function() {
 			return that.db[that.settings.nodeCollection].drop();
 		});
 	},
 	remove: function() {
-
+		if (!this.db)
+			throw Error('Need to connect!');
 	}
 };
